* Fixed some tautologies.
* Don't allow dht trackers to be inserted when dht manager is invalid.


git-svn-id: svn://rakshasa.no/libtorrent/trunk/libtorrent@1271 e378c898-3ddf-0310-93e7-cc216c733640

+++ src/torrent/data/file_list.cc
@@ -570,7 +570,7 @@ FileList::create_chunk_part(FileList::iterator itr, uint64_t offset, uint32_t le
   offset -= (*itr)->offset();
   length = std::min<uint64_t>(length, (*itr)->size_bytes() - offset);
 
-  if (offset < 0)
+  if ((int64_t)offset < 0)
     throw internal_error("FileList::chunk_part(...) caught a negative offset");
 
   // Check that offset != length of file.
